Replace the Log class with std::ostream
Removes the limit that only types known to the Log class can be printed.

Printing of ADDRESSes needs to be cleaned up.  It would be ideal to have
std::ostream << ADDRESS implemented to print them in hex, but it's going
to require more work.  Probaby a new class to replace ADDRESS.

+++ codegen/chllcode.cpp
@@ -21,7 +21,6 @@
 
 #include "boomerang.h"
 #include "exp.h"
-#include "log.h"
 #include "proc.h"
 #include "prog.h"
 #include "signature.h"
@@ -1557,7 +1556,7 @@ CHLLCode::AddProcDec(UserProc *proc, bool open)
 
 	if (parameters.size() > 10 && open) {
 		LOG << "Warning: CHLLCode::AddProcDec: Proc " << proc->getName()
-		    << " has " << (int)parameters.size() << " parameters\n";
+		    << " has " << parameters.size() << " parameters\n";
 	}
 
 	bool first = true;
